Update column text subheader block size.
The column text subheader block size refers to the size of the block which contains compression, creator, name, label, and format information (size = QL - 16|20). For indexing purposes, the text block starts immediately after the subheader's signature.

Previously, we were reading from the correct offset (immediately after the signature) but were reading the column text subheader block size number of bytes. Thus, labels for the last column in the dataset were being cut off as were missing n bytes at the end where n is equal to the number of bytes in between the subheader signature and the start of compression data.